Source Linkedin Ads: apply stream slices
