+++ airbyte-integrations/connectors/source-linkedin-ads/source_linkedin_ads/source.py
@@ -11,9 +11,7 @@ from airbyte_cdk.sources.streams import Stream
 from airbyte_cdk.sources.streams.http.requests_native_auth import Oauth2Authenticator, TokenAuthenticator
 from airbyte_cdk.utils import AirbyteTracedException
 from airbyte_protocol.models import FailureType
-from source_linkedin_ads.streams import (
-    Accounts,
-    AccountUsers,
+from source_linkedin_ads.analytics_streams import (
     AdCampaignAnalytics,
     AdCreativeAnalytics,
     AdImpressionDeviceAnalytics,
@@ -25,11 +23,8 @@ from source_linkedin_ads.streams import (
     AdMemberJobTitleAnalytics,
     AdMemberRegionAnalytics,
     AdMemberSeniorityAnalytics,
-    CampaignGroups,
-    Campaigns,
-    Conversions,
-    Creatives,
 )
+from source_linkedin_ads.streams import Accounts, AccountUsers, CampaignGroups, Campaigns, Conversions, Creatives
 
 logger = logging.getLogger("airbyte")
